I just made a couple minor improvements here and there.
I used the player list rather than the entity list to broadcast a message
to all the players. That way you don't have to go searching through a
load of entities to find a few players (because on most servers, the number
of entities hugely outweighs the number of players).

Also, I got rid of some overrides that were not actually necessary. I
had thought they were necessary because my Eclipse was acting funny because
I had two versions of MinecraftServer.class in my build path: one deobfu-
scated .java file from MCP and one from the minecraft_server.jar.

+++ Corundum/world/Zone.java
@@ -36,9 +36,9 @@ public class Zone {
     }
 
     public float getVolume() {
-        float xLength = this.high.getX() - this.low.getX();
-        float yLength = this.high.getY() - this.low.getY();
-        float zLength = this.high.getZ() - this.low.getZ();
+        float xLength = (float) (this.high.getX() - this.low.getX());
+        float yLength = (float) (this.high.getY() - this.low.getY());
+        float zLength = (float) (this.high.getZ() - this.low.getZ());
 
         return xLength * yLength * zLength;
     }
